Map test now pushes jobReqest to Javascript engine. System still doesn't work.
The scheduling master will scan the database and look for work to
be done, and correctly pass it to a worker for processessing (one
at a time, no coccurent scheduling). This biggest missing thing is
making sure that the correct information is passed to the worker
so that it can find the database and correct stack information.
The next nessacry system may be a name lookup system, so plugins can
request each other by name. Right now the name field is fill out be the
worker code, so manager system needs to create some sort of unique name.