The Javascript engine now has a thread pool.
The JS engine now uses the ExecutorService to do parallel jobs.
It also keeps track of the jobs, so that when the master asks about
them later, it can determine status. Right now those tracked jobs
are never expunged, so the system will keep alocating memory.
Need a system so that the job manager let workers know it is done
a job, and the records can be deleted.