Changing RemusWeb to a RemusNet based interface.
By implementing the RemusNet interface, a webserver can directly
receive data from a RemusWorker (and provide dynamic input to the
worker as well). Of course this system won't actually work until
there is some sort of naming server in place, so that the worker
will know how to contact the webserver rather then the default data
server.