Adding more functionality to command line program.
Note, this path makes a major change, in that now the 'root' pipeline
(the one that lists the other pipelines) is no longer null. That wasn't
cool with thrift, so now it is '@root'.
Working on adding more functionality to command line interface, including
pipeline/instance/applet listing. Also using jline for the console now